requirements.txt
flask==3.0.2
openai==1.12.0
python-dotenv==1.0.1
mysql-connector-python==8.3.0
flask-cors==4.0.0 